requirements.txt
beautifulsoup4==4.12.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0
requests==2.25.1